+++ src/features/side_bar/components/AppNameAndIcon/AppNameAndIcon.tsx
@@ -1,7 +1,7 @@
 import { ActionsBarBTN } from '@konturio/ui-kit';
 import { appConfig } from '~core/app_config';
 import { i18n } from '~core/localization';
-import { transformIconLink } from '~utils/common';
+import { transformIconLink } from '~utils/transformIconLink';
 
 type AppNameAndIconProps = {
   wrapClassName?: string;
