+++ src/utils/common/error.ts
@@ -1,0 +1,8 @@
+export const isErrorWithMessage = (
+  e: unknown,
+): e is {
+  message: string;
+  [key: string]: any;
+} => {
+  return e !== null && typeof e === 'object' && 'message' in e;
+};
