+++ src/utils/common/removeEmpty.ts
@@ -1,0 +1,13 @@
+type NoUndefinedField<T> = {
+  [P in keyof T]-?: Exclude<T[P], null | undefined>;
+};
+
+export const removeEmpty = <T extends Record<string, unknown | null | undefined>>(
+  obj: T,
+): NoUndefinedField<T> =>
+  Object.keys(obj).reduce((acc, key: keyof T) => {
+    if (obj[key] !== undefined && obj[key] !== null) {
+      acc[key] = obj[key] as NoUndefinedField<T>[keyof T];
+    }
+    return acc;
+  }, {} as NoUndefinedField<T>);
