+++ src/utils/common/sorting.ts
@@ -1,0 +1,10 @@
+export const sortByKey =
+  <T extends Record<string, unknown>>(key: string, direction: 'asc' | 'desc') =>
+  (a: T, b: T) => {
+    const aVal = a[key];
+    const bVal = b[key];
+    if (typeof aVal === 'number' && typeof bVal === 'number') {
+      return direction === 'desc' ? bVal - aVal : aVal - bVal;
+    }
+    return 0;
+  };
