+++ src/utils/hooks/useUnlistedRef.ts
@@ -1,0 +1,8 @@
+import { useRef } from 'react';
+
+/* Use it when you want read value in effect without adding it in deps */
+export function useUnlistedRef<T>(value: T) {
+  const valueRef = useRef(value);
+  valueRef.current = value;
+  return valueRef;
+}
